+++ docs/CHANGELOG.md
@@ -2,6 +2,38 @@
 
 The entire project (i.e., the GraphQL API, REST API, and server) use the same version number, and follow [semver](https://semver.org/). Backwards compatibility gets based off of the API (i.e., the GraphQL and REST APIs). For example, if the server has a backward incompatible change such as a DB schema update, but the APIs haven't changed, then the major version doesn't get bumped. Another example is if the GraphQL API hasn't changed, but the format of the HTTP request used to send the GraphQL document has changed, then the major version gets bumped.
 
+## v0.7.0
+
+### Server
+
+- New:
+    - Docker `db` service
+- Updated:
+    - `.env` file
+- Removed:
+    - Docker `auth` service
+    - Docker `auth-db` service
+    - Docker `chat-db` service
+    
+### GraphQL API
+
+- New:
+    - `Name` `scalar`
+    - `Mutation.verifyEmailAddress`
+    - `Mutation.resetPassword`
+    - `Mutation.emailPasswordResetCode`
+- Updated:
+    - `Mutation.updateAccount`
+    - `Mutation.sendEmailAddressVerification` renamed to `Mutation.emailEmailAddressVerification`
+    - `UpdatedAccount` `type`
+    - `AccountData` `interface`
+    - `Account` `type`
+    - `NewContact` `type`
+    - `AccountInput` `input`
+    - `AccountUpdate` `input`
+- Removed:
+    - `Mutation.resetPassword`
+
 ## v0.6.1
 
 ### Server